Generalized approach to stock prediction
The backtest transformer is a generalized approach to stock prediction. Instead of having a program for different data, we have a single program that takes can predict the closing price of a stock. Features are added to the data beforehand from another script so we can further decouple the model from the data.